+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './fieldExperiments'
+export * from './fieldPersonalization'
 export * from './types'
 export * from './utils/flattenSchemaType'