+++ src/types.ts
@@ -18,7 +18,7 @@ export type ExperimentType = {
   variants: VariantType[]
 }
 
-export type FieldPluginConfig = {
+export type ExperimentFieldPluginConfig = {
   fields: (string | FieldDefinition)[]
   experiments: ExperimentType[] | ((client: SanityClient) => Promise<ExperimentType[]>)
   apiVersion?: string
@@ -29,20 +29,38 @@ export type FieldPluginConfig = {
   experimentId?: string
 }
 
+export type PersonalizationFieldPluginConfig = {
+  fields: (string | FieldDefinition)[]
+  segments: VariantType[] | ((client: SanityClient) => Promise<VariantType[]>)
+  apiVersion?: string
+  personalizationNameOverride?: string
+  segmentNameOverride?: string
+  segmentId?: string
+  segmentArrayName?: string
+}
+
 export type VariantPreviewProps = Omit<PreviewProps, 'SchemaType'> & {
   [key: string]: string
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   value: any
 }
 
-export type ExperimentContextProps = Required<FieldPluginConfig> & {
+export type ExperimentContextProps = Required<ExperimentFieldPluginConfig> & {
   experiments: ExperimentType[]
+}
+
+export type PersonalizationContextProps = Required<PersonalizationFieldPluginConfig> & {
+  segments: VariantType[]
 }
 
 export type ArrayInputProps = ArrayOfObjectsInputProps & {
   variantName: string
   variantId: string
   experimentId: string
+}
+export type PersonalizationArrayInputProps = ArrayOfObjectsInputProps & {
+  segmentName: string
+  segmentId: string
 }
 
 export type ObjectFieldWithPath = ObjectField<SchemaType> & {path: Path}